==> 2_merge.R <==
# ================================================
# 1. Split Merge 
# ---------------
# Script to merge the 2 & 3 term document matrices
# generated by the tm application.  The term document
# matrices were processed using subsets of the 
# blogs, news, and twitter files.  Here we merge 
# the individual 2 & 3 term files to create individual
# dataframes named "df4."
# ================================================

library(stringr)
library(dplyr)

for (i in 1:26) {

# Set dir to move files to
dirFile1 <- paste("t",i, "/tdm2.csv", sep="")  
dirFile2 <- paste("t",i, "/tdm3.csv", sep="")   
  
# Read in 2 & 3 Term-Doc Matrices
tdm2 <- read.csv(dirFile1, header = F, as.is = T)
tdm3 <- read.csv(dirFile2, header = F, as.is = T)

# Create empty 2 Term matrix to hold split terms
mRow <- dim(tdm2)[1]
mCol <- 4
m2 <- matrix(nrow = mRow, ncol = mCol)
rm(mRow,mCol)

# Create empty 3 Term matrix to hold split terms
mRow <- dim(tdm3)[1]
mCol <- 4
m3 <- matrix(nrow = mRow, ncol = mCol)
rm(mRow,mCol)

# Split terms and counts of tdm2 into matrix 
m2[,2:3] <- str_split_fixed(tdm2$V1, " ", 2)
m2[,4] <- tdm2$V2
rm(tdm2)

# Split terms and counts of tdm3 into matrix 
m3[,1:3] <- str_split_fixed(tdm3$V1, " ", 3)
m3[,4] <- tdm3$V2
rm(tdm3)

# Remove 1st row of m2 if blank (second column)
if (m2[1,2] == "")
  m2 <- m2[-1,]

# Remove 1st row of m3 if blank
if (m3[1,1] == "")
  m3 <- m3[-1,]

# Create data frame for tdm2
df2 <- data.frame(wrd1=m2[,1],wrd2=m2[,2],wrd3=m2[,3],
                  cnt=as.numeric(m2[,4]),
                  stringsAsFactors = F)
rm(m2)

# Create data frame for tdm3
df3 <- data.frame(wrd1=m3[,1],wrd2=m3[,2], wrd3=m3[,3],
                  cnt=as.numeric(m3[,4]),
                  stringsAsFactors = F)
rm(m3)

# Merge and sort the tdm2 and tdm3
df4 <- rbind(df2,df3) %>%
  arrange(wrd2, wrd1, wrd3)

rm(df2,df3)

# ================================================
# 2. Clean up
# -----------
# Script to clean up some of the problems observed
# in the individual "df4" files.  Words with leading
# dashes, the same word repeated in wrd1 and wrd2 
# are removed here.
# ================================================

# Remove duplicate words in wrd1 and wrd2
col1_2<- (df4$wrd1 == df4$wrd2)
colNAs <- is.na(col1_2)
col1_2[colNAs] <- FALSE
sum(col1_2)
df4 <- df4[!col1_2,]
rm(col1_2,colNAs)

# Remove duplicate words in wrd2 and wrd3
col2_3<- (df4$wrd2 == df4$wrd3)
colNAs <- is.na(col2_3)
col2_3[colNAs] <- FALSE
sum(col2_3)
df4 <- df4[!col2_3,]
rm(col2_3,colNAs)

# Remove dashes : wrd1
sum(grepl("^-", df4$wrd1))
dashed <- grepl("^-", df4$wrd1)
df4 <- df4[!dashed,]
rm(dashed)

# Remove dashes : wrd2
sum(grepl("^-", df4$wrd2))
dashed <- grepl("^-", df4$wrd2)
df4 <- df4[!dashed,]
rm(dashed)

# Remove dashes : wrd3
sum(grepl("^-", df4$wrd3))
dashed <- grepl("^-", df4$wrd3)
df4 <- df4[!dashed,]
rm(dashed)

# Remove dashes : wrd1
sum(grepl("-$", df4$wrd1))
dashed <- grepl("^-", df4$wrd1)
df4 <- df4[!dashed,]
rm(dashed)

# Remove dashes : wrd2
sum(grepl("-$", df4$wrd2))
dashed <- grepl("^-", df4$wrd2)
df4 <- df4[!dashed,]
rm(dashed)

# Remove dashes : wrd3
sum(grepl("-$", df4$wrd3))
dashed <- grepl("^-", df4$wrd3)
df4 <- df4[!dashed,]
rm(dashed)

# ================================================
# 3. Final merge
# --------------
# Script to merge the data frame files processed 
# using subsets of the blogs, news, and twitter 
# files.  Here the "df4" files generated by 
# script #1 and #2 are merged into a single data 
# frame referred to as "df."  The counts of all 
# matching word choices are summed to ensure the
# word with the greatest count is selected.  
# ================================================

df <- as.data.frame(df) 

# Merge and sort the tdm2 and tdm3
df <- rbind(df,df4) %>%
  arrange(wrd2, wrd1, wrd3) %>%
  group_by(wrd2, wrd1,wrd3) %>%
  summarise(cnt=sum(cnt))

rm(df4)

}
