==> 3_reduce.R <==

# ==========================================
# 4. Reduce the data set 
# ----------------------
# Script to reduce the size of the data set 
# by eliminating words where wrd1 and wrd2 
# matched but contained a lower word count than 
# another matching row. This eliminated word
# choices with lower counts.  This script 
# was applied to the individual alphabetically
# ordered data sets generated by the n# 
# scripts.
# ==========================================

library(dplyr)

# Repeat for each data set generated by 
# the n# scripts.

# Get list of each file in dir
rdsFileList <- dir()
rdsFileListCnt <- length(rdsFileList)

# Outer for loop to repeat for each file
for (rdsFileN in 1:rdsFileListCnt) {

# Open rdsFile as "df"  
rdsFile <- rdsFileList[rdsFileN]
df <- readRDS(rdsFile)

# Create a temporary matrix to hold matched rows
dfSize = dim(df)[1]
mSize = (dfSize * 2) + 2
mSzHalf = mSize/2

# Create temp matrix large enough to hold df matches
m <- matrix(numeric(mSize),nrow=mSzHalf,ncol=2)
mRow <- 1

# Set initial booleans
foundMatch <- FALSE
firstMatch <- TRUE

# Create logical vector for rows to delete
tfDel <- logical(dfSize)

# For each row in df
for (i in 2:nrow(df)) {

  # Remove NAs
  if (is.na(df[i,1])) {df[i,1] <- ""} 
  if (is.na(df[i-1,1])) {df[i-1,1] <- ""}
  if (is.na(df[i,2])) {df[i,2] <- ""} 
  if (is.na(df[i-1,2])) {df[i-1,2] <- ""} 
  
  # If last wrd2 same as current wrd2, and,
  # last wrd1 same as current wrd1, 
  # store last row# and cnt cols in m,
  # increment mRow
  
  if (df[i,1] == df[i - 1,1] & df[i,2] == df[i - 1,2]){
    
    # Match was initially false, so prev row is first
    if (foundMatch == FALSE) {
      if (i == 2) {
        firstMatchRow <-  1
      }
      else if (firstMatch) {
        firstMatchRow <- (mRow)
      }
      else{
        firstMatchRow <- (mRow - 1)
      }
    }
    
    # Set match true
    foundMatch <- TRUE
    
    m[mRow,1] <- (i - 1)
    m[mRow,2] <- as.integer(df[(i - 1),4])
    
    mRow <- mRow + 1
    
  } # end if
  
  # If foundMatch is true,
  # but no match found this time,
  # set foundMatch to false, 
  # store last row# and count cols 
  # in m, get the maximum count in m, 
  # get the last row # of m, 
  #   If count of df rows in m < max 
  #   by the threshold val remove the row.
  # reset matrix m, set mRow back to 1
  
  else if (foundMatch == TRUE) {
    foundMatch <- FALSE
    
    m[mRow,1] <- as.integer(i - 1)
    m[mRow,2] <- as.integer(df[(i - 1),4])
    
    # Find last matrix row
    maxMatrixRow <- mRow
    
    # Find max count 
    maxFnd <- max(m[firstMatchRow:maxMatrixRow,2])
    
    # Reduce threshold for deleting row
    if (maxFnd > 500) {
      maxFnd <- maxFnd - 200
    } else if (maxFnd > 100 & maxFnd < 501) {
      maxFnd <- maxFnd - 30
    } else if (maxFnd > 50 & maxFnd < 201) {
      maxFnd <- maxFnd - 10
    } else if (maxFnd > 10 & maxFnd < 51) {
      maxFnd <- maxFnd - 5
    } else if (maxFnd > 2 & maxFnd < 11) {
      maxFnd <- maxFnd - 2
    } else if (maxFnd == 2) {
      maxFnd <- maxFnd - 1  # won't eliminate any
    }  
    
    # Mark rows for deletion
    for (n in firstMatchRow:maxMatrixRow){
      x <- m[n,1]
      if (df[x,4] < maxFnd){
        # Mark row for deletion
        tfDel[x] <- TRUE
        # print(x)
      } # end if
    } # end for
    
    mRow <- mRow + 1
    
  } # end else if
  
} # end for loop

# --------
# Final if
# --------
# if you end on a matched list, find max
if (foundMatch == TRUE){
  m[mRow,1] <- as.integer(i)
  m[mRow,2] <- as.integer(df[i,4])

  # Find last matrix row
  maxMatrixRow <- mRow
    
  # Find max count
  maxFnd <- max(m[firstMatchRow:maxMatrixRow,2])
  
  # Reduce threshold for deleting row
  if (maxFnd > 500) {
    maxFnd <- maxFnd - 200
  } else if (maxFnd > 100 & maxFnd < 501) {
    maxFnd <- maxFnd - 30
  } else if (maxFnd > 50 & maxFnd < 201) {
    maxFnd <- maxFnd - 10
  } else if (maxFnd > 10 & maxFnd < 51) {
    maxFnd <- maxFnd - 5
  } else if (maxFnd > 2 & maxFnd < 11) {
    maxFnd <- maxFnd - 2
  } else if (maxFnd == 2) {
    maxFnd <- maxFnd - 1  # won't eliminate any
  }  
  
  # Mark rows for deletion  
  for (n in firstMatchRow:maxMatrixRow){
    x <- m[n,1]
    if (df[x,4] < maxFnd){
      # Mark row for deletion
      tfDel[x] <- TRUE
      # print(x)
    } # end if
  } # end for loop
  
} # end final if

# Delete rows marked for deletion
df <- df[!tfDel,]

# Sort by 1st, 2nd word; descending cnts
df %>%
  arrange(wrd2, wrd1, desc(cnt), wrd3) 

# Save the reduced size data set
# Overwrite rdsFile
saveRDS(df,rdsFile)

} # end outer for loop

# -------------
# End of Script
# -------------
